Create endpoint that calls the YouTube API
By typing in the ID of the playlist you want the endpoint will return
the all the data. In this case the data is the YouTube title and the
ID of the video itself. This is feature is validated through open docs.

This commit also fixes some of the import issues and removes the testing
code inside of youtube_api.py.